test(ci): split broker security tests
